table/SLR1_FlyAI/requirements.txt
torch
flyai
torchvision
network
numpy
keras
utils
Pillow
